Customer Details Moved to Page 1
